requirements.txt
streamlit
librosa
pandas
fastdtw
scipy
scikit-learn
audio-recorder-streamlit
cloud-sql-python-connector
PyMySQL
bcrypt
google-cloud-storage




